requirements.txt
confluent_kafka==2.2.0
prometheus_client==0.17.1
psycopg2-binary==2.9.6
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.2.1
async-timeout==4.0.2
backcall==0.2.0
backports.functools-lru-cache
brotlipy==0.7.0
certifi
cffi
charset-normalizer
click==8.1.6
cloudpickle
comm==0.1.3
cryptography
cytoolz
dask 
debugpy==1.6.7
decorator==5.1.1
efficientnet
efficientnet-pytorch==0.7.1
entrypoints
exceptiongroup==1.1.2
executing==1.2.0
faiss-cpu==1.7.4
faiss-gpu==1.7.2
fastapi==0.100.1
filelock
fsspec 
gmpy2 
h11==0.14.0
h5py 
idna 
imagecodecs 
imageio 
importlib-metadata==6.8.0
ipykernel==6.24.0
ipython==8.12.2
jedi==0.18.2
Jinja2
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keras
Keras-Applications
Keras-Preprocessing
locket 
MarkupSafe 
matplotlib-inline==0.1.6
mkl-fft==1.3.1
mkl-random 
mkl-service==2.4.0
mpmath==1.2.1
nest-asyncio==1.5.6
networkx
numpy 
packaging==23.1
parso==0.8.3
partd 
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.9.1
pooch 
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2 
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser 
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
pyOpenSSL 
PySocks 
python-dateutil==2.8.2
PyWavelets 
PyYAML 
pyzmq==25.1.0
redis==4.6.0
requests 
scikit-image 
scikit-learn==1.3.0
scipy==1.10.1
six 
sniffio==1.3.0
stack-data==0.6.2
starlette==0.27.0
sympy 
threadpoolctl==3.2.0
tifffile 
toolz 
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
urllib3 
uvicorn==0.23.1
wcwidth==0.2.6
zipp==3.16.2
